Factor out Client and ServerApp from HelloStuff.
Refactor ivm uitbreiden naar meerdere methodes later. main methodes apart van RMI fiunctiona
