+++ javachat/ServerApp.java
@@ -1,0 +1,12 @@
+package javachat;
+
+import javachat.rmi.*;
+
+public class ServerApp {
+
+    RmiRegistry registry;
+
+    public static void main(String args[]) {        
+        RmiRegistry registry = new RmiRegistry();
+    }
+}
