+++ javachat/rmi/RmiRegistry.java
@@ -1,24 +1,20 @@
+package javachat.rmi;
+
 import java.rmi.registry.Registry;
 import java.rmi.AlreadyBoundException;
 import java.rmi.registry.LocateRegistry;
 import java.rmi.server.UnicastRemoteObject;
-        
-public class HelloRmiServer implements HelloRmiInterface {
-        
-    public HelloRmiServer() {}
 
-    public String sayHello() {
-        return "Hello, world! 😀";
-    }
-        
-    public static void main(String args[]) {
-        
+public class RmiRegistry {
+    
+    public RmiRegistry() {
+        Registry registry = LocateRegistry.createRegistry(1099);
         try {
-            HelloRmiServer obj = new HelloRmiServer();
-            HelloRmiInterface skeleton = (HelloRmiInterface) UnicastRemoteObject.exportObject(obj, 0);
 
-            // Bind the remote object's stub in the registry
-            Registry registry = LocateRegistry.createRegistry(1099);
+            // Bind the remote object's stub in the registry.
+            HelloRmiServer helloServer = new HelloRmiServer();
+            HelloRmiInterface skeleton = (HelloRmiInterface) UnicastRemoteObject.exportObject(helloServer, 0);
+
             try {
                 registry.bind(HelloRmiInterface.NAAM, skeleton);
             } catch (AlreadyBoundException e) {
